Move : 파일 이동 및 정리 - request, response 폴더 생성
