+++ src/main/java/com/ybe/tr1ll1on/domain/product/response/AccommodationDetailResponse.java
@@ -1,4 +1,4 @@
-package com.ybe.tr1ll1on.domain.product.dto;
+package com.ybe.tr1ll1on.domain.product.response;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
