+++ src/main/java/com/ybe/tr1ll1on/domain/product/response/AccommodationImageResponse.java
@@ -1,4 +1,4 @@
-package com.ybe.tr1ll1on.domain.product.dto;
+package com.ybe.tr1ll1on.domain.product.response;
 
 import com.ybe.tr1ll1on.domain.accommodation.model.AccommodationImage;
 import lombok.Builder;
