+++ src/test/java/com/ybe/tr1ll1on/domain/product/service/ProductServiceTest.java
@@ -8,10 +8,10 @@ import com.ybe.tr1ll1on.domain.accommodation.model.Accommodation;
 import com.ybe.tr1ll1on.domain.accommodation.model.AccommodationFacility;
 import com.ybe.tr1ll1on.domain.accommodation.model.Category;
 import com.ybe.tr1ll1on.domain.accommodation.repository.AccommodationRepository;
-import com.ybe.tr1ll1on.domain.product.dto.AccommodationDetailResponse;
-import com.ybe.tr1ll1on.domain.product.dto.AccommodationRequest;
-import com.ybe.tr1ll1on.domain.product.dto.ProductFacilityResponse;
-import com.ybe.tr1ll1on.domain.product.dto.ProductResponse;
+import com.ybe.tr1ll1on.domain.product.response.AccommodationDetailResponse;
+import com.ybe.tr1ll1on.domain.product.request.AccommodationRequest;
+import com.ybe.tr1ll1on.domain.product.response.ProductFacilityResponse;
+import com.ybe.tr1ll1on.domain.product.response.ProductResponse;
 import com.ybe.tr1ll1on.domain.product.model.Product;
 import com.ybe.tr1ll1on.domain.product.model.ProductFacility;
 import com.ybe.tr1ll1on.domain.product.repository.ProductInfoPerNightRepository;
@@ -21,7 +21,6 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
